Clean up services and models.
