+++ src/models/coordinates.ts
@@ -1,4 +1,4 @@
-export interface Coordinates {
+export default interface Coordinates {
     lat: number;
     lon: number;
 };
