--- src/services/distance/test.ts
@@ -1,4 +1,0 @@
-import { getDirectionData } from '../direction/directionsApi'
-
-
-console.log(getDirectionData)
